fix: gate RO-broadcast borrow behind borrowModeActive single source of truth
Both borrow read-sites (canonicalize fan-out + scheduler pass
borrowModeEnabled) duplicated the IARA_BROADCAST_OWNERSHIP env lookup and
neither required data-triggered alloc, so priming + join-owns-buffer took the
borrow path even though borrows need data-triggered's distinct per-iteration
buffers (priming reuses one buffer). Collapse both to iara::util::
borrowModeActive() = dataTriggeredActive() && ownership==join-owns-buffer.

Borrow stays default ON (data-triggered + join-owns-buffer); disabling it
(priming, or IARA_BROADCAST_OWNERSHIP=copy-all-but-one) falls back to the copy
path — correct results, just more memory. The scheduler pass now emits the
effective decision to iara_runtime_config.h as IARA_BROADCAST_BORROW so app and
runtime code assert on the compiler's actual choice rather than re-deriving it
from env at runtime.

22-ro-broadcast oracle keys its reuse-count assertion on IARA_BROADCAST_BORROW
(borrow => N reuses/0 copies; copy path => 1 reuse/N-1 copies) and checks 0
mistakes in every mode; matrix widened to both ownership arms. 12/12 green.

Also fix CommonTypes.h u32 typedef (was uint64_t) and mark roadmap T4 done.

+++ include/Iara/Util/CommonTypes.h
@@ -9,7 +9,7 @@ using i8 = int8_t;
 using u8 = uint8_t;
 
 using i32 = int32_t;
-using u32 = uint64_t;
+using u32 = uint32_t;
 
 using i64 = int64_t;
 using u64 = uint64_t;
